Refactor NeuronModels package imports
