+++ NeuronModels/__init__.py
@@ -1,0 +1,3 @@
+from .synaptic_neuron import SynapticNeuron
+
+__all__ = ["SynapticNeuron"]
