+++ NeuronModels/synaptic_neuron.py
@@ -1,461 +1,214 @@
 import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-import os
 
-from synapticNeuronClass import SynapticNeuron  # make sure synaptic_neuron.py is available in the same directory
-# from utils import fft_membrane_potential
+def sigmoid(x: np.array) -> np.array:
+    result = 1 / (1 + np.exp(-x))
+    return result
 
-# def sigmoid(x: np.array) -> np.array:
-#     result = 1 / (1 + np.exp(-x))
-#     return result
+def forward_euler(y, dt, dydt):
+    y_ret = y + dydt*dt
+    return y_ret
 
-# def forward_euler(y, dt, dydt):
-#     y_ret = y + dydt*dt
-#     return y_ret
+class SynapticNeuron:
+    def __init__(
+        self,
+        excitatory_Vin: np.ndarray,
+        inhibitory_Vin: np.ndarray,
+        # Neuron parameters
+        cap: float = 0.82,
+        k: float = 250.0,
+        V0: float = -52.0, 
+        Vs0: float = -50.0, 
+        Vus0: float = -52.0,
+        g_f: float = 1.0,
+        g_s: float = 0.5,
+        g_us: float = 0.015,
+        tau_s: float = 4.3,
+        tau_us: float = 278.0,
+        V_threshold: float = 20.0, 
+        V_peak: float = 20.0,
+        V_reset: float = -45,
+        Vs_reset: float = 7.5,
+        delta_Vus: float = 1.7,
+        # Synaptic parameters
+        Ve0: float = 0,
+        Vi0: float = -90,
+        Ve_threshold: float = -20, 
+        Vi_threshold: float = -20,
+        tau_e: float = 50,
+        tau_i: float = 50,
+        g_syn_e: float = 0.5,
+        g_syn_i: float = 30,
+        I_ext=0.0
+    ):
+        # Neuron parameters
+        self.cap = cap
+        self.k = k
+        self.V0 = V0
+        self.Vs0 = Vs0
+        self.Vus0 = Vus0
+        self.g_f = g_f
+        self.g_s = g_s
+        self.g_us = g_us
+        self.tau_s = tau_s
+        self.tau_us = tau_us
+        self.V_threshold = V_threshold
+        self.V_peak = V_peak
+        self.V_reset = V_reset
+        self.Vs_reset = Vs_reset
+        self.delta_Vus = delta_Vus
 
+        # Synaptic parameters
+        self.Ve0 = Ve0
+        self.Vi0 = Vi0
+        self.Ve_threshold = Ve_threshold
+        self.Vi_threshold = Vi_threshold
+        self.tau_e = tau_e
+        self.tau_i = tau_i
+        self.g_syn_e = g_syn_e
+        self.g_syn_i = g_syn_i
 
-# Baseline simulation without excitatory/inhibitory inputs
-def simulate_neuron_baseline(current_ext, dt, runtime):
-    print("Simulating neuron WITHOUT linked EXCIT/INHIB...")
-    time = np.arange(0, runtime, dt)
-    # numsteps = int(runtime / dt)
+        self.I_ext = I_ext
+        self.excitatory_Vin = np.atleast_1d(excitatory_Vin) if excitatory_Vin is not None else np.array(V0)
+        self.inhibitory_Vin = np.atleast_1d(inhibitory_Vin) if inhibitory_Vin is not None else np.array(V0)
+        self.Se = np.zeros_like(self.excitatory_Vin)
+        self.Si = np.zeros_like(self.inhibitory_Vin)
 
-    neuron = SynapticNeuron(None, None)
+        self.V = V0
+        self.Vs = Vs0
+        self.Vus = Vus0
 
-    excit_ext = np.full_like(time, neuron.V0)
-    inhib_ext = np.full_like(time, neuron.V0)   
+        # Logged data
+        self.Vvalues = []
+        self.Vsvalues = []
+        self.Vusvalues = []
+        self.Sevalues = []
+        self.Sivalues = []
+        self.I_excitatory_values = []
+        self.I_inhibitory_values = []
+
+        # self.I_ext = I_ext
+        # self.excitatory_Vin = excitatory_Vin if excitatory_Vin is not None else Ve0
+        # self.inhibitory_Vin = inhibitory_Vin if inhibitory_Vin is not None else Vi0
+
+    def compute_derivatives(self, log_currents=True):
+        """
+        Se, Si corresponds to the probability of the synapse being open
+        g_synapse = g_syn_e * Se || g_syn_i * Si
+        """
+
+        dVs = self.k * (self.V - self.Vs) / self.tau_s
+        dVus = self.k * (self.V - self.Vus) / self.tau_us
+
+        sigmoid_Ve = sigmoid(40.0 * (self.excitatory_Vin - self.Ve_threshold))
+        sigmoid_Vi = sigmoid(40.0 * (self.inhibitory_Vin - self.Vi_threshold))
+        dSe = self.k * (sigmoid_Ve - self.Se) / self.tau_e
+        dSi = self.k * (sigmoid_Vi - self.Si) / self.tau_i
+
+        I_excitatory = np.sum(self.g_syn_e * self.Se * (self.V - self.Ve0))
+        I_inhibitory = np.sum(self.g_syn_i * self.Si * (self.V - self.Vi0))
+        
+        dV = (
+            self.g_f * (self.V - self.V0)**2
+            - self.g_s * (self.Vs - self.Vs0)**2
+            - self.g_us * (self.Vus - self.Vus0)**2
+            + self.I_ext
+            - I_excitatory
+            - I_inhibitory
+        ) * (self.k / self.cap)
+
+        if log_currents:
+            self.I_excitatory_values.append(I_excitatory)
+            self.I_inhibitory_values.append(I_inhibitory)
+
+        return dV, dVs, dVus, dSe, dSi
     
-    for i, (_, I_ext) in enumerate(zip(time, current_ext)):
-        # Pure excitatory input
-        neuron.update_inputs(I_ext=I_ext, excitatory_Vin=excit_ext[i], inhibitory_Vin=inhib_ext[i])
-        neuron.update_state(dt)
+    def update_inputs(self, I_ext = None, excitatory_Vin = None, inhibitory_Vin = None):
+        if I_ext is not None:
+            self.I_ext = I_ext
+        if excitatory_Vin is not None:
+            self.excitatory_Vin = excitatory_Vin
+        if inhibitory_Vin is not None:
+            self.inhibitory_Vin = inhibitory_Vin
 
-    print(f"Max Se: {max(neuron.Sevalues)}")
-    print(f"Min Se: {min(neuron.Sevalues)}")
-    print(f"Max Si: {max(neuron.Sivalues)}")
-    print(f"Min Si: {min(neuron.Sivalues)}")
+    def update_state(self, dt):
+        dV, dVs, dVus, dSe, dSi = self.compute_derivatives()
+        
+        # First calculate value of V_new 
+        V_new = forward_euler(self.V, dt, dV)
 
-    print(f"Max V: {max(neuron.Vvalues)}")
-    print(f"Min V: {min(neuron.Vvalues)}")
+        if V_new >= self.V_threshold:
+            # Set neuron voltage to fixed V_peak
+            self.V = self.V_peak
+
+            # Recompute new derivatives based on V=V_peak
+            dV_spike, dVs_spike, dVus_spike, dSe_spike, dSi_spike = self.compute_derivatives(log_currents=False)
+
+            # # Update values of Vs, Vus, Se, Si
+            Vs_new = forward_euler(self.Vs, dt, dVs_spike)
+            Vus_new = forward_euler(self.Vus, dt, dVus_spike)
+            Se_new = forward_euler(self.Se, dt, dSe_spike)
+            Si_new = forward_euler(self.Si, dt, dSi_spike)
+            
+            # Append peak value V_peak to Vvalues
+            self.Vvalues.append(self.V)
+
+            # Reset after spiking action
+            self.V = self.V_reset
+            self.Vs = self.Vs_reset
+            self.Vus += self.delta_Vus
+            # self.Se = np.zeros_like(self.excitatory_Vin)
+            # self.Si = np.zeros_like(self.inhibitory_Vin)
+        else:
+            self.V = V_new
+
+            Vs_new = forward_euler(self.Vs, dt, dVs)
+            Vus_new = forward_euler(self.Vus, dt, dVus)
+            Se_new = forward_euler(self.Se, dt, dSe)
+            Si_new = forward_euler(self.Si, dt, dSi)
+
+            self.Vs = Vs_new
+            self.Vus = Vus_new
+            self.Se = Se_new
+            self.Si = Si_new
+
+            self.Vvalues.append(self.V)
+
+        # self.Vsvalues.append(self.Vs)
+        # self.Vusvalues.append(self.Vus)
+        self.Sevalues.append(self.Se)
+        self.Sivalues.append(self.Si)
+
+
+    """
+    def Vs_dot(self):
+        return self.k * (self.V - self.Vs) / self.tau_s
     
-    plt.figure(figsize=(12, 9))
-    plt.subplot(4, 1, 1)
-    plt.plot(time, current_ext, label="Current Input current_ext")
-    plt.title('Current Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('I_ext (mA/nF)')
-    plt.grid()
-    plt.legend()
+    def Vus_dot(self):
+        return self.k * (self.V - self.Vus) / self.tau_us
 
-    plt.subplot(4, 1, 2)
-    plt.plot(time, neuron.Sevalues, label="Se")
-    plt.plot(time, neuron.Sivalues, label="Si")
-    plt.title('Se, Si')
-    plt.xlabel('Time (s)')
-    plt.ylabel('Se, Si')
-    plt.grid()
-    plt.legend()
-
-    plt.subplot(4, 1, 3)
-    plt.plot(time, excit_ext, label="Excitatory Input voltage")
-    plt.plot(time, neuron.Ve_threshold*np.ones_like(time),
-                label="Excit threshold Ve_th",
-                linestyle='dotted',
-                alpha=0.75)
-    plt.plot(time, inhib_ext, label="Inhibitory Input voltage")
-    plt.plot(time, neuron.Vi_threshold*np.ones_like(time),
-                label="Inhib threshold Vi_th",
-                linestyle='dotted',
-                alpha=0.75)
-    plt.title('Excit/Inhib Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-
-    plt.subplot(4, 1, 4)
-    plt.plot(time, neuron.Vvalues, label="Membrane Potential")
-    plt.title('Membrane Potential')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-
-    plt.tight_layout()
-
-    # fft_membrane_potential(neuron, dt)
-
-    plotname = f'SynapticNeuron_Plots/Baseline_{runtime}s_{np.max(current_ext)}.png'
-    print("Saving figure as", plotname)
-    plt.savefig(os.path.join(os.path.dirname(__file__), plotname))
-    # plotname = f'SynapticNeuron_Plots/Baseline_{runtime}s_{np.max(current_ext)}.png'
-    # print("Saving figure as", plotname)
-    # plt.savefig(os.path.join(os.path.dirname(__file__), plotname))
-    # plt.show()
+    #! For half-centre synapse there are 2 inhibitory inputs for each neuron, make it an array of array
+    def Se_dot(self, excitatory_Vin) -> np.ndarray:
+        # Non-linear function to control synaptic gating
+        sigmoid_Ve = self.sigmoid((excitatory_Vin - self.Ve_threshold)*40)
+        return self.k * (sigmoid_Ve - self.Se) / self.tau_e
     
-    plt.close()
-
-    return neuron
-
-def simulate_neuron_excit(current_ext, dt, runtime):
-    print("Simulating neuron with EXCITATORY inputs...")
-    numsteps = int(runtime / dt)
-    time = np.arange(0, runtime, dt)
-
-    neuron = SynapticNeuron(None, None)
+    def Si_dot(self, inhibitory_Vin) -> np.ndarray:
+        # Non-linear function to control synaptic gating
+        sigmoid_Vi = self.sigmoid((inhibitory_Vin - self.Vi_threshold)* 40)
+        return self.k * (sigmoid_Vi - self.Si) / self.tau_i
     
-    # Excitatory input pulse
-    amplitude = 100
-    peak_time = 0.2
-    decay_time = 0.1
+    def V_dot(self, I_ext):
+        
+        Ie = self.g_syn_e * self.Se * (self.Ve0 - self.V)
+        Ii = self.g_syn_i * self.Si * (self.Vi0 - self.V)
+        
+        return self.k * (I_ext + np.sum(Ie) + np.sum(Ii) + self.g_f*((self.V-self.V0)**2) - self.g_s*((self.Vs-self.Vs0)**2) - self.g_us*((self.Vus-self.Vus0)**2)) / self.cap
     
-    excit_ext = -52 + amplitude * np.exp(-((time - peak_time - 0.3*runtime)**2)/(2*decay_time**2))
-    inhib_ext = np.full_like(time, neuron.V0)
-
-    excit_ext = np.array(excit_ext)
-    inhib_ext = np.array(inhib_ext)
-
-    print(f"Max excit: {max(excit_ext)}")
-    print(f"Min excit: {min(excit_ext)}")
-    print(f"Max inhib: {max(inhib_ext)}")
-    print(f"Min inhib: {min(inhib_ext)}")
-    
-    for i, (t, I_ext) in enumerate(zip(time, current_ext)):
-        # Pure excitatory input
-        neuron.update_inputs(I_ext=I_ext, excitatory_Vin=excit_ext[i], inhibitory_Vin=inhib_ext[i])
-        neuron.update_state(dt)
-
-    print(f"Max Se: {max(neuron.Sevalues)}")
-    print(f"Min Se: {min(neuron.Sevalues)}")
-    print(f"Max Si: {max(neuron.Sivalues)}")
-    print(f"Min Si: {min(neuron.Sivalues)}")
-
-    plt.figure(figsize=(12, 9))
-    plt.subplot(4, 1, 1)
-    plt.plot(time, current_ext, label="Current Input current_ext")
-    plt.title('Current Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('I_ext (mA/nF)')
-    plt.grid()
-    plt.legend()
-
-    plt.subplot(4, 1, 2)
-    plt.plot(time, neuron.Sevalues, label="Se")
-    plt.title('Se')
-    plt.xlabel('Time (s)')
-    plt.ylabel('Se')
-    plt.grid()
-    plt.legend()
-
-
-    plt.subplot(4, 1, 3)
-    plt.plot(time, excit_ext, label="Excitatory Input voltage")
-    plt.plot(time, neuron.Ve_threshold*np.ones_like(time), 
-             label="Excit threshold Ve_th", 
-             linestyle='dotted', 
-             alpha=0.75)
-    plt.plot(time, inhib_ext, label="Inhibitory Input voltage")
-    plt.plot(time, neuron.Vi_threshold*np.ones_like(time), 
-             label="Inhib threshold Vi_th", 
-             linestyle='dotted', 
-             alpha=0.75)
-    plt.title('Excitatory Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-
-
-    plt.subplot(4, 1, 4)
-    plt.plot(time, neuron.Vvalues, label="Membrane Potential")
-    plt.title('Membrane Potential')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-    
-    plt.tight_layout()
-    plotname = f'SynapticNeuron_Plots/Excit_{decay_time}decay_{runtime}s_{np.max(current_ext)}.png'
-    print("Saving figure as", plotname)
-    plt.savefig(os.path.join(os.path.dirname(__file__), plotname))
-
-    # plt.show()
-    plt.close()
- 
-    plt.figure(figsize=(12, 6))
-    plt.plot(time, neuron.I_excitatory_values, label="I_excitatory")
-    plt.title('I_excitatory')
-    plt.xlabel('Time (s)')
-    plt.ylabel('I_excitatory (mA/nF)')
-    plt.grid()
-    plt.legend()
-    # plt.show()
-    plt.savefig(os.path.join(os.path.dirname(__file__), f'SynapticNeuron_Plots/Excit_{decay_time}decay_{runtime}s_{np.max(current_ext)}_current.png'))
-
-    plt.close()
-
-    return neuron
-
-def simulate_neuron_inhib(current_ext, dt, runtime):
-    print("Simulating neuron with INHIBITORY inputs...")
-    numsteps = int(runtime / dt)
-    time = np.arange(0, runtime, dt)
-    
-    neuron = SynapticNeuron(None, None)
-    
-    # Inhibitory input pulse
-    amplitude = 100
-    peak_time = 0.2
-    decay_time = 0.1
-
-    excit_ext = np.full_like(time, neuron.V0)
-    inhib_ext = -52 + amplitude * np.exp(-((time - peak_time - 0.3*runtime)**2)/(2*decay_time**2))
-
-    excit_ext = np.array(excit_ext)
-    inhib_ext = np.array(inhib_ext)
-
-    print(f"Max excit: {max(excit_ext)}")
-    print(f"Min excit: {min(excit_ext)}")
-    print(f"Max inhib: {max(inhib_ext)}")
-    print(f"Min inhib: {min(inhib_ext)}")
-
-    for i, (t, I_ext) in enumerate(zip(time, current_ext)):
-        # Pure inhibitory input
-        neuron.update_inputs(I_ext=I_ext, excitatory_Vin=excit_ext[i], inhibitory_Vin=inhib_ext[i])
-        neuron.update_state(dt)
-
-    print(f"Max Se: {max(neuron.Sevalues)}")
-    print(f"Min Se: {min(neuron.Sevalues)}")
-    print(f"Max Si: {max(neuron.Sivalues)}")
-    print(f"Min Si: {min(neuron.Sivalues)}")
-
-    plt.figure(figsize=(12, 9))
-    plt.subplot(4, 1, 1)
-    plt.plot(time, current_ext, label="Current Input current_ext")
-    plt.title('Current Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('I_ext (mA/nF)')
-    plt.grid()
-    plt.legend()
-    
-    plt.subplot(4, 1, 2)
-    plt.plot(time, neuron.Sivalues, label="Si")
-    plt.title('Si')
-    plt.xlabel('Time (s)')
-    plt.ylabel('Si')
-    plt.grid()
-    plt.legend()
-
-
-    plt.subplot(4, 1, 3)
-    plt.plot(time, excit_ext, label="Excitatory Input voltage")
-    plt.plot(time, neuron.Ve_threshold*np.ones_like(time), 
-             label="Excit threshold Ve_th", linestyle='dotted',
-             alpha=0.75)
-    plt.plot(time, inhib_ext, label="Inhibitory Input voltage")
-    plt.plot(time, neuron.Vi_threshold*np.ones_like(time), 
-             label="Inhib threshold Vi_th", 
-             linestyle='dotted',
-             alpha=0.75)
-    plt.title('Inhibitory Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-
-
-    plt.subplot(4, 1, 4)
-    plt.plot(time, neuron.Vvalues, label="Membrane Potential")
-    plt.title('Membrane Potential')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-    
-    plt.tight_layout()
-    plotname = f'SynapticNeuron_Plots/Inhib_{decay_time}decay_{runtime}s_{np.max(current_ext)}.png'
-    print("Saving figure as", plotname)
-    plt.savefig(os.path.join(os.path.dirname(__file__), plotname))
-
-    # plt.show()
-    plt.close()
-
-    plt.figure(figsize=(12, 6))
-    plt.plot(time, neuron.I_inhibitory_values, label="I_inhibitory")
-    plt.title('I_inhibitory')
-    plt.xlabel('Time (s)')
-    plt.ylabel('I_inhibitory (mA/nF)')
-    plt.grid()
-    plt.legend()
-    # plt.show()
-    plt.savefig(os.path.join(os.path.dirname(__file__), f'SynapticNeuron_Plots/Inhib_{decay_time}decay_{runtime}s_{np.max(current_ext)}_current.png'))
-
-    plt.close()
-
-    return neuron
-
-def simulate_neuron(current_ext, dt, runtime):
-    print("Simulating neuron with BOTH excitatory and inhibitory inputs...")
-    numsteps = int(runtime / dt)
-    time = np.arange(0, runtime, dt)
-    
-    neuron = SynapticNeuron(None, None)
-    
-    # Inhibitory input pulse
-    amplitude = 100
-    peak_time = 0.2
-    decay_time = 0.01
-
-    excit_ext = -52 + amplitude * np.exp(-((time - peak_time - 0.2*runtime)**2)/(2*decay_time**2))
-
-    inhib_ext = -52 + amplitude * np.exp(-((time - peak_time - 0.6*runtime)**2)/(2*decay_time**2))
-
-
-    excit_ext = np.array(excit_ext)
-    inhib_ext = np.array(inhib_ext)
-
-    # print(f"Max excit: {max(excit_ext)}")
-    # print(f"Min excit: {min(excit_ext)}")
-    # print(f"Max inhib: {max(inhib_ext)}")
-    # print(f"Min inhib: {min(inhib_ext)}")
-
-    for i, (t, I_ext) in enumerate(zip(time, current_ext)):
-        # Pure excitatory input
-        neuron.update_inputs(I_ext=I_ext, excitatory_Vin=excit_ext[i], inhibitory_Vin=inhib_ext[i])
-        neuron.update_state(dt)
-
-    # print(f"Max Se: {max(neuron.Sevalues)}")
-    # print(f"Min Se: {min(neuron.Sevalues)}")
-    # print(f"Max Si: {max(neuron.Sivalues)}")
-    # print(f"Min Si: {min(neuron.Sivalues)}")
-
-    plt.figure(figsize=(12, 9))
-    # plt.subplot(5, 1, 1)
-    plt.subplot(4, 1, 1)
-    plt.plot(time, current_ext, label="Current Input current_ext")
-    plt.title('Current Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('I_ext (mA/nF)')
-    plt.grid()
-    plt.legend()
-    
-    plt.subplot(4, 1, 2)
-    plt.plot(time, neuron.Sevalues, label="Se")
-    plt.plot(time, neuron.Sivalues, label="Si")
-    plt.title('Se, Si')
-    plt.xlabel('Time (s)')
-    plt.ylabel('Se, Si')
-    plt.grid()
-    plt.legend()
-
-    plt.subplot(4, 1, 3)
-    plt.plot(time, excit_ext, label="Excitatory Input voltage")
-    plt.plot(time, neuron.Ve_threshold*np.ones_like(time), 
-             label="Excit threshold Ve_th", 
-             linestyle='dotted',
-             alpha=0.75)
-    plt.plot(time, inhib_ext, label="Inhibitory Input voltage")
-    plt.plot(time, neuron.Vi_threshold*np.ones_like(time), 
-             label="Inhib threshold Vi_th", 
-             linestyle='dotted',
-             alpha=0.75)
-    plt.title('External Excit/Inhib Input')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-
-
-    plt.subplot(4, 1, 4)
-    plt.plot(time, neuron.Vvalues, label="Membrane Potential")
-    plt.title('Membrane Potential')
-    plt.xlabel('Time (s)')
-    plt.ylabel('V (mV)')
-    plt.grid()
-    plt.legend()
-    
-    plt.tight_layout()
-    plotname = f'SynapticNeuron_Plots/Both_{decay_time}decay_{runtime}s_{np.max(current_ext)}.png'
-    print("Saving figure as", plotname)
-    plt.savefig(os.path.join(os.path.dirname(__file__), plotname))
-
-    # plt.show()
-    plt.close()
-
-
-    plt.figure(figsize=(12, 6))
-    plt.plot(time, neuron.I_excitatory_values, label="I_excitatory", color='red')
-    plt.plot(time, neuron.I_inhibitory_values, label="I_inhibitory", color='blue')
-    plt.title('I_excitatory & I_inhibitory')
-    plt.xlabel('Time (s)')
-    plt.ylabel('I (mA/nF)')
-    plt.grid()
-    plt.legend()
-    # plt.show()
-
-    plotname = f'SynapticNeuron_Plots/Both_{decay_time}decay_{runtime}s_{np.max(current_ext)}_current.png'
-    plt.savefig(os.path.join(os.path.dirname(__file__), plotname))
-    plt.close()
-
-    return neuron
-
-    # sigmoid_time = sigmoid(time)
-    # plt.figure(figsize=(12, 6))
-    # plt.plot(time, sigmoid_time, label="Sigmoid")
-    # plt.title('Sigmoid')
-    # plt.xlabel('Time (s)')
-    # plt.ylabel('Sigmoid')
-    # plt.grid()
-    # plt.legend()
-    # plt.savefig(os.path.join(os.path.dirname(__file__), 'synaptic_neuron_sigmoid.png'))
-    # plt.show()
-
-
-def main():
-    dt = 5e-5
-    # runtimes = [5.0, 10.0, 15.0]
-    runtimes = [5.0]
-    amplitude = 5
-
-    for runtime in runtimes:
-        numsteps = int(runtime / dt)
-        time = np.arange(0, runtime, dt)
-        # Constant current input
-        current_ext = np.zeros(numsteps)
-        start_time = int(0.5 / dt)
-        current_ext[start_time:] = amplitude
-
-        "Simulate baseline"
-        neuron_baseline = simulate_neuron_baseline(current_ext, dt, runtime)
-        csvpath = os.path.join(os.path.dirname(__file__), f'SynapticNeuron_Plots/{runtime}s_I={amplitude}_baseline.csv')
-        data = pd.DataFrame({'time': time, 'V': neuron_baseline.Vvalues})
-        data.to_csv(csvpath, index=False)
-        print("Saved baseline data to", csvpath)
-
-        "Simulate excitatory input"
-        neuron_excit = simulate_neuron_excit(current_ext, dt, runtime)
-        csvpath = os.path.join(os.path.dirname(__file__), f'SynapticNeuron_Plots/{runtime}s_I={amplitude}_excit.csv')
-        data = pd.DataFrame({'time': time, 'V': neuron_excit.Vvalues})
-        data.to_csv(csvpath, index=False)
-        print("Saved excitatory data to", csvpath)
-
-        """Simulate inhibitory input"""
-        neuron_inhib = simulate_neuron_inhib(current_ext, dt, runtime)
-        csvpath = os.path.join(os.path.dirname(__file__), f'SynapticNeuron_Plots/{runtime}s_I={amplitude}_inhib.csv')
-        data = pd.DataFrame({'time': time, 'V': neuron_inhib.Vvalues})
-        data.to_csv(csvpath, index=False)
-        print("Saved inhibitory data to", csvpath)
-
-        """Simulate both excitatory and inhibitory inputs"""
-        neuron_EI = simulate_neuron(current_ext, dt, runtime)
-        csvpath = os.path.join(os.path.dirname(__file__), f'SynapticNeuron_Plots/{runtime}s_I={amplitude}_EI.csv')
-        data = pd.DataFrame({'time': time, 'V': neuron_EI.Vvalues})
-        data.to_csv(csvpath, index=False)
-        print("Saved EI data to", csvpath)
-
-    
-
-
-if __name__ == "__main__":
-    main()
+    def update(self, dt, I_ext):
+        new_Se = self.Se + dt * self.Se_dot(self.excitatory_Vin)
+        new_Si = self.Si + dt * self.Si_dot(self.inhibitory_Vin)
+        new_Vs = self.Vs + dt * self.Vs_dot()
+        new_Vus = self.Vus + dt * self.Vus_dot()
+        new_V = self.V + dt * self.V_dot(I_ext)
+        return new_Se, new_Si, new_Vs, new_Vus, new_V
+    """
